refactor : presentation 파일구조 변경
