+++ test/saved_recipes_screen_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:recipe_app/recipe_app/mock/mock_recipe_data_impl.dart';
-import 'package:recipe_app/recipe_app/presentation/home/saved_recipes_screen.dart';
-import 'package:recipe_app/recipe_app/presentation/home/saved_recipes_view_model.dart';
+import 'package:recipe_app/recipe_app/presentation/saved_recipes_screen/saved_recipes_screen.dart';
+import 'package:recipe_app/recipe_app/presentation/saved_recipes_screen/saved_recipes_view_model.dart';
 import 'package:recipe_app/recipe_app/repository/recipe_repository.dart';
 import 'package:recipe_app/recipe_app/repository/recipe_repository_impl.dart';
