+++ test/splash_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:recipe_app/recipe_app/mock/mock_recipe_data_impl.dart';
-import 'package:recipe_app/recipe_app/presentation/home/splash_screen.dart';
+import 'package:recipe_app/recipe_app/presentation/splash_screen/splash_screen.dart';
 import 'package:recipe_app/recipe_app/repository/recipe_repository.dart';
 import 'package:recipe_app/recipe_app/repository/recipe_repository_impl.dart';
